app/src/main/java/com/example/dicodingstoryapp/data/viewmodel/HomeViewModel.kt
package com.example.dicodingstoryapp.data.viewmodel

import androidx.lifecycle.LiveData
import androidx.lifecycle.ViewModel
import androidx.lifecycle.ViewModelProvider
import androidx.lifecycle.viewModelScope
import androidx.paging.PagingData
import androidx.paging.cachedIn
import com.example.dicodingstoryapp.data.response.ListStoryItem
import com.example.dicodingstoryapp.paging.StoryRepository

class HomeViewModel(storyRepository: StoryRepository) : ViewModel() {

    val story: LiveData<PagingData<ListStoryItem>> =
        storyRepository.getStoryStream().cachedIn(viewModelScope)

    class HomeViewModelFactory(private val storyRepository: StoryRepository) :
        ViewModelProvider.Factory {
        override fun <T : ViewModel> create(modelClass: Class<T>): T {
            if (modelClass.isAssignableFrom(HomeViewModel::class.java)) {
                return HomeViewModel(storyRepository) as T
            }
            throw IllegalArgumentException("Unknown ViewModel class")
        }
    }
}